Extract out SSO Login, other tidies
