+++ connect/amazon/iamidentitycenter/client/src/main/kotlin/org/http4k/connect/amazon/iamidentitycenter/PKCES256Generator.kt
@@ -12,12 +12,13 @@ object PKCES256Generator {
     fun generate(): Pair<PKCEChallenge, PKCECodeVerifier> {
         val randomBytes = ByteArray(32).apply(secureRandom::nextBytes)
         val codeVerifier = Base64.getUrlEncoder().withoutPadding().encodeToString(randomBytes)
-        val challenge =
-            Base64.getUrlEncoder().withoutPadding().encodeToString(MessageDigest.getInstance("SHA-256").apply {
-                update(codeVerifier.toByteArray(Charsets.UTF_8))
-            }.digest())
+        val challenge = Base64.getUrlEncoder().withoutPadding().encodeToString(
+            MessageDigest
+                .getInstance("SHA-256")
+                .apply { update(codeVerifier.toByteArray(Charsets.UTF_8)) }
+                .digest()
+        )
         return PKCEChallenge.of(challenge) to PKCECodeVerifier.of(codeVerifier)
     }
-
 }
